IMba Chnges

+++ Assets/Script/Weapon/type of Weapon/WeaponParametersSO.cs
@@ -1,18 +1,13 @@
-using System.Collections;
-using System.Collections.Generic;
-using Unity.VisualScripting;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "NewWeaponParameters", menuName = "WeaponParametersSO")]
 public class WeaponParametersSO : ScriptableObject
 {
-
-        
         public float damage;
         public float range;
         public float impactForce;
         public float fireRate;
-        public float magazine;
-        public float magazineCapacity;
+        public int magazine;
+        public int magazineCapacity;
         public float _audioclips;
 }
